+++ src/Switchyard.Worker/IInventoryOutboxDispatchCycle.cs
@@ -1,0 +1,9 @@
+using Switchyard.Messaging;
+
+namespace Switchyard.Worker;
+
+public interface IInventoryOutboxDispatchCycle
+{
+    Task<OutboxDispatchResult> DispatchAsync(
+        CancellationToken cancellationToken);
+}